Separate selector and selection packages
