+++ bundles/framework/tools.vitruv.framework.vsum/META-INF/MANIFEST.MF
@@ -13,8 +13,8 @@ Require-Bundle: org.apache.log4j,
  org.eclipse.emf.ecore.xmi,
  edu.kit.ipd.sdq.activextendannotations
 Export-Package: tools.vitruv.framework.vsum,
- tools.vitruv.framework.vsum.models,
  tools.vitruv.framework.vsum.internal;x-internal:=true,
+ tools.vitruv.framework.vsum.models,
  tools.vitruv.framework.vsum.views,
- tools.vitruv.framework.vsum.views.selection
+ tools.vitruv.framework.vsum.views.selectors
 Bundle-Vendor: vitruv.tools
