+++ bundles/framework/tools.vitruv.framework.vsum/src/tools/vitruv/framework/vsum/views/selectors/TreeBasedViewSelector.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.vsum.views.selection
+package tools.vitruv.framework.vsum.views.selectors
 
 import org.eclipse.emf.ecore.EObject
 import tools.vitruv.framework.vsum.views.ViewSelector
